# openapi.yaml
openapi: 3.0.3
info:
  title: WC-Hub API
  version: 1.0.0
  description: |
    WC-Hub é um control plane multi-cloud para gerenciamento unificado de infraestrutura.
    
    **Recursos principais:**
    - Gestão de VMs/containers (Proxmox, Docker, Kubernetes, OCI)
    - Infraestrutura como código (Terraform)
    - Integrations (GitHub, Cloudflare)
    - RBAC granular com policy enforcement
    - Audit logging completo
    - Terminal SSH via WebSocket
    
    **Autenticação:**
    - Session-based authentication com cookies HTTPOnly
    - TOTP 2FA opcional
    - Policy enforcement para ações críticas (X-Confirmation + X-TOTP-Code headers)
    
  contact:
    name: WC-Hub Development Team
  license:
    name: Proprietary
servers:
  - url: https://hub.example.com
    description: Production
  - url: http://localhost:8080
    description: Development
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/proxmox/nodes/{node}/{kind}/{vmid}/snapshots:
    get:
      summary: GET /api/v1/proxmox/nodes/{node}/{kind}/{vmid}/snapshots
      operationId: get_proxmox_nodes_node_kind_vmid_snapshots
      tags: [proxmox]
      parameters:
        - name: node
          in: path
          required: true
          schema:
            type: string
        - name: kind
          in: path
          required: true
          schema:
            type: string
        - name: vmid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operação concluída
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: POST /api/v1/proxmox/nodes/{node}/{kind}/{vmid}/snapshots
      operationId: post_proxmox_nodes_node_kind_vmid_snapshots
      tags: [proxmox]
      parameters:
        - name: node
          in: path
          required: true
          schema:
            type: string
        - name: kind
          in: path
          required: true
          schema:
            type: string
        - name: vmid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operação concluída
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/proxmox/nodes/{node}/{kind}/{vmid}/snapshots/{name}:
    delete:
      summary: DELETE /api/v1/proxmox/nodes/{node}/{kind}/{vmid}/snapshots/{name}
      operationId: delete_proxmox_nodes_node_kind_vmid_snapshots_name
      tags: [proxmox]
      parameters:
        - name: node
          in: path
          required: true
          schema:
            type: string
        - name: kind
          in: path
          required: true
          schema:
            type: string
        - name: vmid
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operação concluída
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Acesso negado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'